FIXED: HEADER DETAILS AND LOCALIZED FAULT HANDLERS

+++ include/idt.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <kernel/fos.h>
 
-void idt_set_gate(uint8_t num, uint32_t base, uint16_t sel, uint8_t flags);
+void idt_set_gate(uint8_t num, uint64_t base, uint16_t sel, uint8_t flags);
 
 void idt_install();
